split working files to make the _init_ clean
